+++ crates/miden-testing/tests/agglayer/mod.rs
@@ -5,7 +5,7 @@ mod config_bridge;
 mod faucet_helpers;
 mod global_index;
 mod leaf_utils;
-mod mmr_frontier;
+mod merkle_tree_frontier;
 mod solidity_miden_address_conversion;
 pub mod test_utils;
 mod update_ger;